+++ lib/compiler/src/lib.rs
@@ -84,10 +84,13 @@ pub use crate::target::{
     PointerWidth, Target, Triple,
 };
 #[cfg(feature = "translator")]
+#[allow(deprecated)]
+pub use crate::translator::to_wasm_error;
+#[cfg(feature = "translator")]
 pub use crate::translator::{
-    to_wasm_error, translate_module, wptype_to_type, FunctionBodyData, FunctionMiddleware,
-    MiddlewareBinaryReader, MiddlewareReaderState, ModuleEnvironment, ModuleInfoTranslation,
-    ModuleMiddleware, ModuleMiddlewareChain, ModuleTranslationState,
+    translate_module, wptype_to_type, FunctionBodyData, FunctionMiddleware, MiddlewareBinaryReader,
+    MiddlewareReaderState, ModuleEnvironment, ModuleInfoTranslation, ModuleMiddleware,
+    ModuleMiddlewareChain, ModuleTranslationState,
 };
 pub use crate::trap::TrapInformation;
 pub use crate::unwind::CompiledFunctionUnwindInfo;